Cleanup; Add CHIP SPA form
