+++ src/services/ui/src/router.tsx
@@ -65,7 +65,8 @@ export const router = createBrowserRouter([
         path: ROUTES.CHIP_ELIGIBILITY_LANDING,
         element: <P.CHIPEligibilityLandingPage />,
       },
-      { path: ROUTES.MEDICAID_SPA_FORM, element: <P.SampleFormPage /> },
+      { path: ROUTES.MEDICAID_SPA_FORM, element: <P.MedicaidSpaForm /> },
+      { path: ROUTES.CHIP_SPA_FORM, element: <P.ChipSpaForm /> },
       { path: "/form", element: <P.ExampleForm /> },
     ],
     loader: rootLoader(queryClient),
